pyproject.toml
[project]
name = "snowmin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "wjdecorte", email = "jdecorte@decorteindustries.com" }
]
requires-python = ">=3.14"
dependencies = [
    "click>=8.1.7",
    "snowflake-connector-python>=3.12.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "jinja2>=3.1.4",
    "rich>=13.9.0",
    "pyyaml>=6.0.2",
    "colorama>=0.4.6",
    "cryptography>=41.0.0",
]


[project.scripts]
snowmin = "snowmin.cli:cli"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.14",
    "pytest-click",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["uv_build>=0.9.10,<0.10.0"]
build-backend = "uv_build"
